+++ Runtime/Scripts/IContainer.cs
@@ -19,7 +19,7 @@ namespace ExpressoBits.Inventory
         /// <param name="item">T Item to be added</param>
         /// <param name="amount">Amount to be added</param>
         /// <returns>Amount not added in operation</returns>
-        public byte Add(T item, byte amount);
+        public ushort Add(T item, ushort amount);
 
         /// <summary>
         /// Remove item from index
@@ -27,7 +27,7 @@ namespace ExpressoBits.Inventory
         /// <param name="index">Slot index to be removed</param>
         /// <param name="amount">Amount to be removed</param>
         /// <returns>Amount not removed in operation</returns>
-        public byte RemoveInIndex(int index, byte amount);
+        public ushort RemoveInIndex(int index, ushort amount);
 
         /// <summary>
         /// Remove T item with a quantity and returns values that were not remove.
@@ -35,7 +35,7 @@ namespace ExpressoBits.Inventory
         /// <param name="item">T Item to be removed</param>
         /// <param name="amount">Amount to be removed</param>
         /// <returns>Amount not removed in operation</returns>
-        public byte Remove(T item, byte amount);
+        public ushort Remove(T item, ushort amount);
 
         /// <summary>
         /// Clear container
@@ -55,6 +55,6 @@ namespace ExpressoBits.Inventory
         /// <param name="item"></param>
         /// <param name="amount"></param>
         /// <returns>Return true if container contain T item with amount</returns>
-        public bool Has(T item, byte amount);
+        public bool Has(T item, ushort amount);
     }
 }
