+++ Runtime/Scripts/ISlot.cs
@@ -10,19 +10,19 @@ namespace ExpressoBits.Inventory
         /// Return item id
         /// </summary>
         /// <value></value>
-        public byte ItemID { get; }
+        public ushort ItemID { get; }
 
         /// <summary>
         /// Return value of itens in slot
         /// </summary>
         /// <value></value>
-        public byte Amount  { get; }
+        public ushort Amount  { get; }
 
         /// <summary>
         /// Returns the maximum stack value
         /// </summary>
         /// <value></value>
-        public byte MaxStack { get; }
+        public ushort MaxStack { get; }
 
         /// <summary>
         /// Return true if amount is less than stack
@@ -40,20 +40,20 @@ namespace ExpressoBits.Inventory
         /// Remaining amount to fill the stack
         /// </summary>
         /// <value>Remaining value</value>
-        public byte Remaining { get; }
+        public ushort Remaining { get; }
 
         /// <summary>
         /// Adds item to stack and returns quantity that was not successfully added.
         /// </summary>
         /// <param name="count">Amount to be added</param>
         /// <returns>Unadded quantity</returns>
-        public byte Add(byte count);
+        public ushort Add(ushort count);
 
         /// <summary>
         /// Remove item from stack and return quantity that was not successfully removed.
         /// </summary>
         /// <param name="count"></param>
         /// <returns></returns>
-        public byte Remove(byte count);
+        public ushort Remove(ushort count);
     }
 }
