+++ Runtime/Scripts/Item.cs
@@ -5,34 +5,34 @@ namespace ExpressoBits.Inventory
     [CreateAssetMenu(fileName = "Item", menuName = "Expresso Bits/Inventory/Item")]
     public class Item : ScriptableObject, IItem
     {
-        public byte ID => id;
+        public ushort ID => id;
         public string Name => name;
         public string Description => description;
         public Sprite Icon => icon;
         public float Weight => weight;
-        public byte MaxStack => maxStack;
+        public ushort MaxStack => maxStack;
         public ItemObject ItemObjectPrefab => itemObjectPrefab;
         public Category Category => category;
 
-        [SerializeField] private byte id = 0;
+        [SerializeField] private ushort id = 0;
         [SerializeField, TextArea] private string description;
         [SerializeField] private Sprite icon;
         [SerializeField, Min(0.01f)] private float weight = 0.1f;
-        [SerializeField, Min(1)] private byte maxStack = 64;
+        [SerializeField, Min(1)] private ushort maxStack = 64;
         [SerializeField] private ItemObject itemObjectPrefab;
         [SerializeField] private Category category;
 
-        internal void Setup(byte id)
+        internal void Setup(ushort id)
         {
             this.id = id;
         }
 
-        public static implicit operator byte(Item item)
+        public static implicit operator ushort(Item item)
         {
             return item.ID;
         }
 
-        public static implicit operator Item(byte id)
+        public static implicit operator Item(ushort id)
         {
             Items[] items = Resources.FindObjectsOfTypeAll<Items>();
             if(items.Length > 0)
